* Remove default instance of ConfigValidateFacade
