+++ dubbo-common/src/main/java/org/apache/dubbo/config/ConfigValidateFacade.java
@@ -16,7 +16,6 @@
  */
 package org.apache.dubbo.config;
 
-import org.apache.dubbo.common.deploy.ApplicationDeployListener;
 import org.apache.dubbo.common.extension.ExtensionLoader;
 import org.apache.dubbo.common.logger.Logger;
 import org.apache.dubbo.common.logger.LoggerFactory;
@@ -27,10 +26,9 @@ import org.apache.dubbo.rpc.model.ScopeModel;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicReference;
 
 @SuppressWarnings({"rawtypes", "unchecked"})
-public class ConfigValidateFacade implements ConfigValidator, ApplicationDeployListener {
+public class ConfigValidateFacade implements ConfigValidator{
 
     private static final Logger LOGGER = LoggerFactory.getErrorTypeAwareLogger(ConfigValidateFacade.class);
 
@@ -41,18 +39,12 @@ public class ConfigValidateFacade implements ConfigValidator, ApplicationDeployL
      */
     private static boolean enableValidate = true;
 
-    /**
-     * For test
-     */
-    private static final AtomicReference<ConfigValidateFacade> DEFAULT_INSTANCE = new AtomicReference<>();
-
     public ConfigValidateFacade(ScopeModel scopeModel) {
         if(scopeModel != null) {
             ExtensionLoader<ConfigValidator> extensionLoader = scopeModel.getExtensionLoader(ConfigValidator.class);
             if (extensionLoader != null) {
                 this.validators = extensionLoader.getActivateExtensions();
                 this.validators.forEach(scopeModel.getBeanFactory()::registerBean);
-                DEFAULT_INSTANCE.set(this);
             } else {
                 this.validators = Collections.emptyList();
             }
@@ -105,39 +97,18 @@ public class ConfigValidateFacade implements ConfigValidator, ApplicationDeployL
         return true;
     }
 
-    @Override
-    public void onStarting(ApplicationModel scopeModel) {
-        ConfigValidateFacade validateFacade = new ConfigValidateFacade(scopeModel);
-        DEFAULT_INSTANCE.compareAndSet(null,validateFacade);
-        scopeModel.getBeanFactory().registerBean(validateFacade);
-    }
-
-    @Override
-    public void onStarted(ApplicationModel scopeModel) {}
-
-    @Override
-    public void onInitialize(ApplicationModel scopeModel) {}
-
-    @Override
-    public void onStopping(ApplicationModel scopeModel) {}
-
-    @Override
-    public void onStopped(ApplicationModel scopeModel) {}
-
-    @Override
-    public void onFailure(ApplicationModel scopeModel, Throwable cause) {}
-
-
     /**
      * For test
      */
+    @Deprecated
     public static ConfigValidateFacade getDefaultInstance() {
-        return DEFAULT_INSTANCE.updateAndGet(configValidateFacade -> configValidateFacade == null ? new ConfigValidateFacade(null) : configValidateFacade);
+        return ApplicationModel.defaultModel().getBeanFactory().getOrRegisterBean(ConfigValidateFacade.class);
     }
 
     /**
      * For test
      */
+    @Deprecated
     public static void setEnableValidate(boolean enable){
         enableValidate = enable;
     }
