+++ dubbo-config/dubbo-config-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.config.deploy.lifecycle.application.ApplicationLifecycle
